Automatically tags from query string; reversed PROVIDE/SEEK
After discussion with Dietmar about topics raised in issue #20:

1.  PROVIDE and SEEK are now reverse.  SEEK is always the side with a
radius (the person who is prepared to move).  This could be a waiter
seeking a job anywhere in London, or hungry person in search of a
hamburger.  PROVIDE is always the fixed location with no radius (the
person providing a job or a hamburger).

2.  Instead of the device sending a list of tags, it simply sends a
query string written by a human, which includes Twitter-style
hash-tags.  "#waiter and #bartender seeks work".  In future we might
also extract non-hash-prefixed words if they match a pre-existing tag,
but for now there must be at least one tag with a hash.