Google push notifications, version -1.
"Tickle" notifications only for now, with no payload.  Need to convert
this quick and dirty "speedy-notifier" script into a proper daemon
that stays up and does exponential back-off on errors etc etc... but
generally this appears to be working...  Here is what Google tells me:

{'failure': 1,
 'canonical_ids': 0,
 'success': 0,
 'multicast_id': -1,
 'results': [{'error': 'InvalidRegistration'}]}

Which I think just says that "1234" isn't a phone.  Yeah.