Now returning synchronisation results in JSON format.
As discussed with MM over icecream in the sun.  This avoids the
difficulty of parsing raw SQL scripts, since we now know that Android
SQLite wants to receive the statements one by one, but some of these
statements could be multi-line.  It also makes the synchronisation
results more like every other result the server sends, and this will
become important shortly because we're planning to include "free"
synchronisation results with other results, like create_search (watch
this space).