No more 'tags' or # signs.
Now, we just take all the words (so # has no special meaning, it's
just ignored because it's a non-word, and can therefore separate words
like any other punctuation or whitespace).  Each of these candidate
tags is converted to lower case.  Then we ignore the ones that are
configured as IGNORED or DELETED.  We reject searches that include a
tag that is BANNED.  We only index on the ones that are ACTIVE.  I
stole an initial list of words to ignore from a document linked from
Wikipedia.