Allow messages to be sent supplying a single recipient's fingerprint.
As requested by Mike, you can send either an existing crew ID, or a
fingerprint.  If you send a fingerprint, the server will try to find
an existing crew that has only the sender and the recipient in it, and
send the message in that crew; if it fails to find such a crew it will
create one.