Resolve issue #29 since MAGUIRE asked so nicely.
The 'metadata' element now contains some extra elements like this:

{
    "status": "OK",
    "operation": "incremental",
    "sql": [],
    "message_type": "synchronise_response",
    "metadata": [
        {
            "timeline": 1      <--- the timeline from your request
        },
        {
            "old_sequence": 1  <--- the sequence from your request
        },
        {
            "new_sequence": 1  <--- the new sequence number
        }
    ]
}

I claim this is pointless because the first two values were provided
by the caller so could hardly be news, and the third value is the new
sequence number *after* applying this change set, which can hardly be
useful if you are trying to compute whether the change set should be
applied.  (The change set should be applied only if old_sequence
matches the sequence in the device's database, but that is information
the client already had, since it was provided in the request).

Or I'm confused about the problem.  But let's see if this helps :-)